backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
paapi5-python-sdk
python-amazon-sp-api
